pyproject.toml
[project]
name = "fastapi-mongo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "motor>=3.6.0",
    "odmantic>=1.0.2",
    "python-dotenv>=1.0.1",
]
